+++ CHANGELOG.md
@@ -9,6 +9,9 @@ their feed angles. The feed angles that are stored on the object have their
 feed angles normalized to be between 0 and pi/2.
 
 ### Fixed
+- A bug in `utils.io.ms.read_ms_history` where reading the table caused an IndexError
+due to the "APP_PARAMS" and "CLI_COMMAND" columns being populated in a non-standard
+fashion.
 - A bug in UVBeam's CST reader that caused the bandpass array to be incorrectly
 set to 0 except for the first frequency.
 - A bug in `UVParameter.get_from_form` and `UVParameter.set_from_form` that caused